cp-eras-ui: iced 0.14, per-corner cuts, G2i gate, traces and bars
The pre-work for converting the sixteen traces and four bars into
iced, done in the order that means the Rust is written once.

iced 0.13 -> 0.14.0, iced_layershell 0.13.7 -> 0.19.1. One deviation
from upstream defaults: 0.14 turned `web-colors` (sRGB blending) on
and it drains the kitsch and neokitsch blooms and thins every glyph,
95.7% on the kitsch screens; Cargo.toml restates the default set
without it. The rest is forced: `application(boot, update, view)`,
`Widget::update` for `on_event`, zero-sized `Space` children now
dropped by Row/Column (~56 sites, caught only by the goldens),
`Pixels: From<u16>` gone, canvas text align_x/align_y, scrollable Id
and auto_scroll, `Subscription::run_with` wanting a Hash handle, and
layershell trading `remove_id` for `window::Event::Closed`. tiny-skia
and png stay single-copy. Unverified until a switch: the bar as a
live layer surface. Also new on 0.19.1 and untried: `NewPopUp` grabs,
so the tray menu no longer has to be an output-sized overlay.

`Corners` becomes four `Cut`s (Square | Chamfer {x, y} | Round) with
the era's `Corner` supplying the default, because the neokitsch bar
cell is r3 on three corners and a 10x7 chamfer on the fourth, which no
single field says. Pure refactor: 21 captures byte-identical, seven
tests build the bar shapes and read them back through `span_at`.

`scripts/render.sh` is the golden matrix's recipe outside the sandbox,
7s a capture, and G2i (`fidelity_check.sh --implementation`) diffs a
design SVG against that capture as shape inventories with a verdict.
Bars fail at 11/5/35/7% matched area, logins at 0%: the starting line.
`extract_spec.py`'s ink grid needed the canvas divisible by 45.

Twelve new traces (login, mailbox, store x four eras) and the four
dashboard traces improved against their sources; all sixteen pass
their G1i gate. The four bars redrawn per era from those traces, each
with an IMPLEMENTATION DELTA block for the conversion.

All 21 goldens re-taken on the 0.14 build; matrix 21/21 at 99.9.

+++ home/common/pkgs/cp-eras-ui/src/widgets/text.rs
@@ -9,7 +9,7 @@ use iced::widget::{text, Text};
 use iced::Color;
 
 fn base<'a>(content: impl IntoFragment<'a>, size: u16, color: Color) -> Text<'a> {
-    text(content).size(size).color(color)
+    text(content).size(f32::from(size)).color(color)
 }
 
 pub fn title<'a>(style: &Style, content: impl IntoFragment<'a>) -> Text<'a> {
